Move component properties and property values to separate files
